feature: implement firebase integration, added demo home page with fetched products
