+++ firestore.rules
@@ -1,0 +1,8 @@
+   rules_version = '2';
+   service cloud.firestore {
+     match /databases/{database}/documents {
+       match /{document=**} {
+         allow read, write: if true; // Allow all reads and writes
+       }
+     }
+   }
