--- page.jsx
@@ -1,30 +1,0 @@
-"use client";
-import { useEffect, useState } from 'react';
-import { db } from './firebaseConfig';
-import { collection, getDocs } from 'firebase/firestore';
-
-
-export default function Home() {
-  const [items, setItems] = useState([]);
-  useEffect(() => {
-    const fetchData = async () => {
-      const querySnapshot = await getDocs(collection(db, "products"));
-      const data = querySnapshot.docs.map(doc => ({ id: doc.id, ...doc.data() }));
-      setItems(data);
-    };
-
-    fetchData();
-  }, []);
-
-  return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-        <ul>
-          {items.map(item => (
-            <li key={item.id}>{item.name}</li>
-          ))}
-        </ul>
-      </main>
-    </div>
-  );
-}
